:Add ability to skip initialization
Why do we need this change?
=======================
If a external db exists and the nominatimdb is initialized we don't want
to reinitialize it since the upstream code does a drop of the database
before starting. As such we should have the ability to skip anything
other than starting up the application tier

What effects does this change have?
=======================
* Add's support for a run only mode.
* initializer will do either the restore/init and then the runtime
  container is always just a runtime